back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "matplotlib>=3.10.0",
    "numpy==1.26.4",
    "opencv-python>=4.11.0.86",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "setuptools>=75.8.0",
    "sqlacodegen-v2>=0.1.4",
    "sqlalchemy>=2.0.37",
    "supabase>=2.13.0",
    "tqdm>=4.67.1",
]
